--- src/core/database.ts
@@ -1,917 +1,0 @@
-import { mkdirSync } from "node:fs";
-import { dirname } from "node:path";
-import {
-  type DuckDBConnection,
-  DuckDBInstance,
-  listValue,
-} from "@duckdb/node-api";
-import { config } from "../config/index.js";
-import type { VideoData } from "../types/models/video.js";
-import { AsyncMutex } from "../utils/asyncMutex.js";
-import { logger } from "../utils/logger.js";
-
-/**
- * Discovered user data for tracking new UP主
- */
-interface DiscoveredUserData {
-  userId: bigint;
-  userName: string;
-  fans: number;
-  source: "following" | "recommendation";
-}
-
-/**
- * User statistics update
- */
-interface UserStatsUpdate {
-  videosSeen?: number;
-  videosFiltered?: number;
-  fans?: number;
-  userName?: string;
-}
-
-/**
- * Recommendation data
- */
-interface RecommendationData {
-  videoBvid: string;
-  recommendedByBvid: string;
-  recommendCount: number;
-  recommendOrder: number;
-  firstSeen: Date;
-  lastSeen: Date;
-}
-
-/**
- * User data with statistics
- */
-interface UserData {
-  userId: bigint;
-  userName: string;
-  fans: number;
-  videosSeen: number;
-  videosFiltered: number;
-  filterPassRate: number;
-  discoveredFrom: "following" | "recommendation";
-  discoveredAt: Date;
-  isFollowing: boolean;
-  lastUpdated: Date;
-}
-
-/**
- * Database statistics
- */
-interface DatabaseStats {
-  processedVideosCount: number;
-  forwardDynamicsCount: number;
-  recommendationsCount: number;
-  discoveredUsersCount: number;
-  filteredVideosCount: number;
-}
-
-/**
- * DuckDB database manager - singleton pattern
- * Manages database connection, schema initialization, and CRUD operations
- * Thread-safe through mutex locking
- */
-export class Database {
-  private static instance: Database | null = null;
-  private duckDBInstance: DuckDBInstance | null = null;
-  private connection: DuckDBConnection | null = null;
-  private mutex = new AsyncMutex();
-
-  private constructor() {}
-
-  /**
-   * Get the singleton instance
-   */
-  public static getInstance(): Database {
-    if (!Database.instance) {
-      Database.instance = new Database();
-    }
-    return Database.instance;
-  }
-
-  /**
-   * Initialize the database connection and schema
-   */
-  public async init(path: string = config.database.path): Promise<void> {
-    if (this.connection) {
-      logger.warn("Database already initialized");
-      return;
-    }
-
-    logger.info(`Initializing DuckDB at ${path}`);
-
-    try {
-      // Ensure the database directory exists
-      const dbDir = dirname(path);
-      mkdirSync(dbDir, { recursive: true });
-      logger.debug(`Database directory created/verified: ${dbDir}`);
-
-      // Create or connect to DuckDB instance
-      this.duckDBInstance = await DuckDBInstance.create(path);
-      this.connection = await this.duckDBInstance.connect();
-
-      // Initialize schema
-      await this.initializeSchema();
-
-      logger.info("DuckDB initialized successfully");
-    } catch (error) {
-      logger.error("Failed to initialize DuckDB:", error);
-      throw error;
-    }
-  }
-
-  /**
-   * Execute a database operation with mutex protection
-   */
-  private async withMutex<T>(operation: () => Promise<T>): Promise<T> {
-    const release = await this.mutex.acquire();
-    try {
-      return await operation();
-    } finally {
-      release();
-    }
-  }
-
-  /**
-   * Initialize database schema with all required tables
-   */
-  private async initializeSchema(): Promise<void> {
-    if (!this.connection) {
-      throw new Error("Database not initialized");
-    }
-
-    logger.info("Initializing database schema");
-
-    // Create processed_videos table
-    await this.connection.run(`
-      CREATE TABLE IF NOT EXISTS processed_videos (
-        aid BIGINT PRIMARY KEY,
-        bvid VARCHAR UNIQUE NOT NULL,
-        pubdate BIGINT,
-        title VARCHAR,
-        description TEXT,
-        tag TEXT,
-        pic VARCHAR,
-        type_id INTEGER,
-        user_id BIGINT,
-        is_filtered BOOLEAN NOT NULL,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-      )
-    `);
-
-    // Add new columns if they don't exist (schema migration)
-    const newColumns = [
-      "ALTER TABLE processed_videos ADD COLUMN IF NOT EXISTS staff BIGINT[]",
-      "ALTER TABLE processed_videos ADD COLUMN IF NOT EXISTS tid_v2 INTEGER",
-      "ALTER TABLE processed_videos ADD COLUMN IF NOT EXISTS dynamic TEXT",
-      "ALTER TABLE processed_videos ADD COLUMN IF NOT EXISTS tag_new VARCHAR[]",
-      "ALTER TABLE processed_videos ADD COLUMN IF NOT EXISTS participle VARCHAR[]",
-      "ALTER TABLE processed_videos ADD COLUMN IF NOT EXISTS ctime BIGINT",
-      "ALTER TABLE processed_videos ADD COLUMN IF NOT EXISTS is_deleted BOOLEAN DEFAULT FALSE",
-      "ALTER TABLE processed_videos ADD COLUMN IF NOT EXISTS copyright INTEGER",
-      "ALTER TABLE processed_videos ADD COLUMN IF NOT EXISTS extras JSON",
-      "ALTER TABLE processed_videos ADD COLUMN IF NOT EXISTS notes JSON",
-    ];
-    for (const sql of newColumns) {
-      await this.connection.run(sql);
-    }
-
-    // Migrate JSON array columns to native array types if needed
-    await this.migrateArrayColumns();
-
-    // Create indexes for processed_videos
-    await this.connection.run(`
-      CREATE INDEX IF NOT EXISTS idx_processed_bvid 
-      ON processed_videos(bvid)
-    `);
-
-    await this.connection.run(`
-      CREATE INDEX IF NOT EXISTS idx_processed_user 
-      ON processed_videos(user_id)
-    `);
-
-    await this.connection.run(`
-      CREATE INDEX IF NOT EXISTS idx_processed_filtered 
-      ON processed_videos(is_filtered)
-    `);
-
-    // Create forward_dynamics table
-    await this.connection.run(`
-      CREATE TABLE IF NOT EXISTS forward_dynamics (
-        forward_dynamic_id BIGINT PRIMARY KEY,
-        original_bvid VARCHAR NOT NULL,
-        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-      )
-    `);
-
-    // Create index for forward_dynamics
-    await this.connection.run(`
-      CREATE INDEX IF NOT EXISTS idx_forward_bvid 
-      ON forward_dynamics(original_bvid)
-    `);
-
-    // Create recommendations table
-    await this.connection.run(`
-      CREATE TABLE IF NOT EXISTS recommendations (
-        video_bvid VARCHAR,
-        recommended_by_bvid VARCHAR,
-        recommend_count INTEGER DEFAULT 1,
-        recommend_order INTEGER,
-        first_seen TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        last_seen TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        PRIMARY KEY (video_bvid, recommended_by_bvid)
-      )
-    `);
-
-    // Create indexes for recommendations
-    await this.connection.run(`
-      CREATE INDEX IF NOT EXISTS idx_rec_video 
-      ON recommendations(video_bvid)
-    `);
-
-    await this.connection.run(`
-      CREATE INDEX IF NOT EXISTS idx_rec_count 
-      ON recommendations(recommend_count DESC)
-    `);
-
-    // Create discovered_users table
-    await this.connection.run(`
-      CREATE TABLE IF NOT EXISTS discovered_users (
-        user_id BIGINT PRIMARY KEY,
-        user_name VARCHAR,
-        fans INTEGER DEFAULT 0,
-        videos_seen INTEGER DEFAULT 0,
-        videos_filtered INTEGER DEFAULT 0,
-        filter_pass_rate REAL DEFAULT 0.0,
-        discovered_from VARCHAR,
-        discovered_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        is_following BOOLEAN DEFAULT FALSE,
-        last_updated TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-      )
-    `);
-
-    // Create indexes for discovered_users
-    await this.connection.run(`
-      CREATE INDEX IF NOT EXISTS idx_user_source 
-      ON discovered_users(discovered_from)
-    `);
-
-    await this.connection.run(`
-      CREATE INDEX IF NOT EXISTS idx_user_rate 
-      ON discovered_users(filter_pass_rate DESC)
-    `);
-
-    await this.connection.run(`
-      CREATE INDEX IF NOT EXISTS idx_user_fans 
-      ON discovered_users(fans DESC)
-    `);
-
-    logger.info("Database schema initialized");
-  }
-
-  /**
-   * Migrate JSON array columns to native array types if they exist as JSON.
-   * Uses CREATE TABLE AS SELECT approach for efficient bulk migration.
-   */
-  private async migrateArrayColumns(): Promise<void> {
-    if (!this.connection) {
-      throw new Error("Database not initialized");
-    }
-
-    // Check if migration is needed by inspecting column types
-    const columnsReader = await this.connection.runAndReadAll(`
-      SELECT column_name, data_type 
-      FROM information_schema.columns 
-      WHERE table_name = 'processed_videos' 
-        AND column_name IN ('staff', 'tag_new', 'participle')
-    `);
-
-    const columns = columnsReader.getRowObjects();
-    const needsMigration = columns.some(
-      (col) => (col.data_type as string).toUpperCase() === "JSON",
-    );
-
-    if (!needsMigration) {
-      logger.debug(
-        "Array columns already using native types, skipping migration",
-      );
-      return;
-    }
-
-    logger.info("Migrating JSON array columns to native array types...");
-
-    // Create new table with correct types using EXCLUDE + type cast
-    await this.connection.run(`
-      CREATE TABLE processed_videos_new AS
-      SELECT 
-        * EXCLUDE (staff, tag_new, participle),
-        staff::BIGINT[] AS staff,
-        tag_new::VARCHAR[] AS tag_new,
-        participle::VARCHAR[] AS participle
-      FROM processed_videos
-    `);
-
-    // Drop old table and rename new one
-    await this.connection.run("DROP TABLE processed_videos");
-    await this.connection.run(
-      "ALTER TABLE processed_videos_new RENAME TO processed_videos",
-    );
-
-    // Rebuild indexes
-    await this.connection.run(
-      "CREATE INDEX idx_processed_bvid ON processed_videos(bvid)",
-    );
-    await this.connection.run(
-      "CREATE INDEX idx_processed_user ON processed_videos(user_id)",
-    );
-    await this.connection.run(
-      "CREATE INDEX idx_processed_filtered ON processed_videos(is_filtered)",
-    );
-
-    logger.info("Array column migration completed");
-  }
-
-  /**
-   * Check if a video has been processed
-   */
-  public async hasProcessedVideo(bvid: string): Promise<boolean> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      const reader = await this.connection.runAndReadAll(
-        "SELECT COUNT(*) as count FROM processed_videos WHERE bvid = $1",
-        { 1: bvid },
-      );
-
-      const rows = reader.getRows();
-      return (rows[0]?.[0] as number) > 0;
-    });
-  }
-
-  /**
-   * Check if a video has been processed by ID (AID or BVID)
-   */
-  public async hasProcessedVideoById(
-    id: string | number | bigint,
-  ): Promise<boolean> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      const isBvid = typeof id === "string" && id.startsWith("BV");
-      const sql = isBvid
-        ? "SELECT COUNT(*) as count FROM processed_videos WHERE bvid = $1"
-        : "SELECT COUNT(*) as count FROM processed_videos WHERE aid = $1";
-
-      const param = isBvid ? id : BigInt(id);
-
-      const reader = await this.connection.runAndReadAll(sql, { 1: param });
-
-      const rows = reader.getRows();
-      return (rows[0]?.[0] as number) > 0;
-    });
-  }
-
-  /**
-   * Get all processed video IDs of a specific type (aid or bvid)
-   */
-  public async getAllProcessedIds(type: "aid" | "bvid"): Promise<Set<string>> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      const column = type === "aid" ? "aid" : "bvid";
-      // Select only the specific column to minimize data transfer
-      const reader = await this.connection.runAndReadAll(
-        `SELECT ${column} FROM processed_videos`,
-      );
-
-      const rows = reader.getRows();
-      const ids = new Set<string>();
-
-      for (const row of rows) {
-        if (row[0] !== null && row[0] !== undefined) {
-          ids.add(row[0].toString());
-        }
-      }
-
-      return ids;
-    });
-  }
-
-  /**
-   * Mark a video as processed
-   */
-  public async markVideoProcessed(
-    video: VideoData,
-    filtered: boolean,
-  ): Promise<void> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      await this.connection.run(
-        `
-      INSERT INTO processed_videos 
-        (aid, bvid, pubdate, title, description, tag, pic, type_id, user_id, is_filtered, 
-         staff, tid_v2, dynamic, tag_new, participle, ctime, is_deleted, copyright, extras, notes, updated_at)
-      VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, 
-              $11, $12, $13, $14, $15, $16, $17, $18, $19, $20, NOW())
-      ON CONFLICT (bvid) DO UPDATE SET
-        aid = EXCLUDED.aid,
-        pubdate = EXCLUDED.pubdate,
-        title = EXCLUDED.title,
-        description = EXCLUDED.description,
-        tag = EXCLUDED.tag,
-        pic = EXCLUDED.pic,
-        type_id = EXCLUDED.type_id,
-        user_id = EXCLUDED.user_id,
-        is_filtered = EXCLUDED.is_filtered,
-        staff = EXCLUDED.staff,
-        tid_v2 = EXCLUDED.tid_v2,
-        dynamic = EXCLUDED.dynamic,
-        tag_new = EXCLUDED.tag_new,
-        participle = EXCLUDED.participle,
-        ctime = EXCLUDED.ctime,
-        is_deleted = EXCLUDED.is_deleted,
-        copyright = EXCLUDED.copyright,
-        extras = EXCLUDED.extras,
-        notes = EXCLUDED.notes,
-        updated_at = NOW()
-    `,
-        {
-          1: BigInt(video.aid),
-          2: video.bvid,
-          3: video.pubdate,
-          4: video.title,
-          5: video.description,
-          6: video.tag,
-          7: video.pic,
-          8: video.type_id,
-          9: BigInt(video.user_id),
-          10: filtered,
-          11: video.staff ? listValue(video.staff) : null,
-          12: video.tid_v2 ?? null,
-          13: video.dynamic ?? null,
-          14: video.tag_new ? listValue(video.tag_new) : null,
-          15: video.participle ? listValue(video.participle) : null,
-          16: video.ctime ?? null,
-          17: video.is_deleted ?? false,
-          18: video.copyright ?? null,
-          19: video.extras ? JSON.stringify(video.extras) : null,
-          20: video.notes ? JSON.stringify(video.notes) : null,
-        },
-      );
-    });
-  }
-
-  /**
-   * Get processed videos
-   */
-  public async getProcessedVideos(
-    limit?: number,
-    where?: string,
-  ): Promise<VideoData[]> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      let sql = "SELECT * FROM processed_videos";
-
-      if (where) {
-        sql += ` WHERE ${where}`;
-      }
-
-      sql += " ORDER BY created_at DESC";
-
-      if (limit) {
-        sql += ` LIMIT ${limit}`;
-      }
-
-      const reader = await this.connection.runAndReadAll(sql);
-      const rows = reader.getRowObjects();
-
-      return rows.map((row) => ({
-        aid: row.aid as bigint,
-        bvid: row.bvid as string,
-        pubdate: row.pubdate as number,
-        title: row.title as string,
-        description: row.description as string,
-        tag: row.tag as string,
-        pic: row.pic as string,
-        type_id: row.type_id as number,
-        user_id: row.user_id as bigint,
-        staff: (row.staff as bigint[] | null) ?? undefined,
-        tid_v2: row.tid_v2 as number | undefined,
-        dynamic: row.dynamic as string | undefined,
-        tag_new: (row.tag_new as string[] | null) ?? undefined,
-        participle: (row.participle as string[] | null) ?? undefined,
-        ctime: row.ctime as number | undefined,
-        is_deleted: row.is_deleted as boolean | undefined,
-        copyright: row.copyright as number | undefined,
-        extras: row.extras ? JSON.parse(row.extras as string) : undefined,
-        notes: row.notes ? JSON.parse(row.notes as string) : undefined,
-      }));
-    });
-  }
-
-  /**
-   * Get list of bvids only (lightweight, for batch processing)
-   */
-  public async getBvidList(where?: string): Promise<string[]> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      let sql = "SELECT bvid FROM processed_videos";
-      if (where) {
-        sql += ` WHERE ${where}`;
-      }
-      sql += " ORDER BY created_at DESC";
-
-      const reader = await this.connection.runAndReadAll(sql);
-      const rows = reader.getRows();
-      return rows.map((row) => row[0] as string);
-    });
-  }
-
-  /**
-   * Get cached forward dynamic BVID
-   */
-  public async getCachedForwardBvid(dynamicId: string): Promise<string | null> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      const reader = await this.connection.runAndReadAll(
-        "SELECT original_bvid FROM forward_dynamics WHERE forward_dynamic_id = $1",
-        { 1: BigInt(dynamicId) },
-      );
-
-      const rows = reader.getRows();
-      return rows.length > 0 ? (rows[0]?.[0] as string) : null;
-    });
-  }
-
-  /**
-   * Cache forward dynamic relationship
-   */
-  public async cacheForward(dynamicId: string, bvid: string): Promise<void> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      await this.connection.run(
-        `
-      INSERT INTO forward_dynamics 
-        (forward_dynamic_id, original_bvid)
-      VALUES ($1, $2)
-      ON CONFLICT (forward_dynamic_id) DO UPDATE SET
-        original_bvid = EXCLUDED.original_bvid
-    `,
-        {
-          1: BigInt(dynamicId),
-          2: bvid,
-        },
-      );
-    });
-  }
-
-  /**
-   * Track recommendation relationship
-   */
-  public async trackRecommendation(
-    videoBvid: string,
-    recommendedByBvid: string,
-    order: number,
-  ): Promise<void> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      // Check if recommendation already exists
-      const reader = await this.connection.runAndReadAll(
-        `SELECT recommend_count FROM recommendations 
-         WHERE video_bvid = $1 AND recommended_by_bvid = $2`,
-        { 1: videoBvid, 2: recommendedByBvid },
-      );
-
-      const rows = reader.getRows();
-
-      if (rows.length > 0) {
-        // Update existing recommendation
-        const currentCount = rows[0]?.[0] as number;
-        await this.connection.run(
-          `UPDATE recommendations 
-           SET recommend_count = $1, last_seen = CURRENT_TIMESTAMP, recommend_order = $2
-           WHERE video_bvid = $3 AND recommended_by_bvid = $4`,
-          {
-            1: currentCount + 1,
-            2: order,
-            3: videoBvid,
-            4: recommendedByBvid,
-          },
-        );
-      } else {
-        // Insert new recommendation
-        await this.connection.run(
-          `INSERT INTO recommendations 
-           (video_bvid, recommended_by_bvid, recommend_count, recommend_order)
-           VALUES ($1, $2, $3, $4)`,
-          {
-            1: videoBvid,
-            2: recommendedByBvid,
-            3: 1,
-            4: order,
-          },
-        );
-      }
-    });
-  }
-
-  /**
-   * Get top recommended videos
-   */
-  public async getTopRecommendedVideos(
-    limit: number,
-  ): Promise<RecommendationData[]> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      const reader = await this.connection.runAndReadAll(
-        `SELECT * FROM recommendations 
-         ORDER BY recommend_count DESC 
-         LIMIT $1`,
-        { 1: limit },
-      );
-
-      const rows = reader.getRowObjects();
-
-      return rows.map((row) => ({
-        videoBvid: row.video_bvid as string,
-        recommendedByBvid: row.recommended_by_bvid as string,
-        recommendCount: row.recommend_count as number,
-        recommendOrder: row.recommend_order as number,
-        firstSeen: new Date(row.first_seen as string),
-        lastSeen: new Date(row.last_seen as string),
-      }));
-    });
-  }
-
-  /**
-   * Check if a user exists in the database
-   */
-  public async hasUser(userId: bigint): Promise<boolean> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      const reader = await this.connection.runAndReadAll(
-        "SELECT COUNT(*) as count FROM discovered_users WHERE user_id = $1",
-        { 1: userId },
-      );
-
-      const rows = reader.getRows();
-      return (rows[0]?.[0] as number) > 0;
-    });
-  }
-
-  /**
-   * Add a discovered user
-   */
-  public async addDiscoveredUser(user: DiscoveredUserData): Promise<void> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      await this.connection.run(
-        `INSERT INTO discovered_users 
-         (user_id, user_name, fans, discovered_from, videos_seen, videos_filtered, filter_pass_rate)
-         VALUES ($1, $2, $3, $4, 0, 0, 0.0)
-         ON CONFLICT (user_id) DO UPDATE SET
-           user_name = EXCLUDED.user_name,
-           fans = EXCLUDED.fans`,
-        {
-          1: user.userId,
-          2: user.userName,
-          3: user.fans,
-          4: user.source,
-        },
-      );
-    });
-  }
-
-  /**
-   * Update user statistics
-   */
-  public async updateUserStats(
-    userId: bigint,
-    stats: UserStatsUpdate,
-  ): Promise<void> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      // Build update query dynamically
-      const updates: string[] = [];
-      const params: Record<string, bigint | number | string> = { userId };
-
-      if (stats.videosSeen !== undefined) {
-        updates.push("videos_seen = videos_seen + $videosSeen");
-        params.videosSeen = stats.videosSeen;
-      }
-
-      if (stats.videosFiltered !== undefined) {
-        updates.push("videos_filtered = videos_filtered + $videosFiltered");
-        params.videosFiltered = stats.videosFiltered;
-      }
-
-      if (stats.fans !== undefined) {
-        updates.push("fans = $fans");
-        params.fans = stats.fans;
-      }
-
-      if (stats.userName !== undefined) {
-        updates.push("user_name = $userName");
-        params.userName = stats.userName;
-      }
-
-      // Calculate filter pass rate based on updated values
-      let filterPassRateCalc = "filter_pass_rate";
-      if (
-        stats.videosSeen !== undefined ||
-        stats.videosFiltered !== undefined
-      ) {
-        filterPassRateCalc =
-          "CASE WHEN (videos_seen" +
-          (stats.videosSeen !== undefined ? " + $videosSeen" : "") +
-          ") > 0 THEN CAST((videos_filtered" +
-          (stats.videosFiltered !== undefined ? " + $videosFiltered" : "") +
-          ") AS REAL) / (videos_seen" +
-          (stats.videosSeen !== undefined ? " + $videosSeen" : "") +
-          ") ELSE 0.0 END";
-      }
-      updates.push(`filter_pass_rate = ${filterPassRateCalc}`);
-      updates.push("last_updated = NOW()");
-
-      if (updates.length > 0) {
-        await this.connection.run(
-          `UPDATE discovered_users SET ${updates.join(
-            ", ",
-          )} WHERE user_id = $userId`,
-          params,
-        );
-      }
-    });
-  }
-
-  /**
-   * Get top discovered users
-   */
-  public async getTopDiscoveredUsers(
-    orderBy: "filter_pass_rate" | "fans",
-    limit: number,
-  ): Promise<UserData[]> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      const reader = await this.connection.runAndReadAll(
-        `SELECT * FROM discovered_users 
-         ORDER BY ${orderBy} DESC 
-         LIMIT $1`,
-        { 1: limit },
-      );
-
-      const rows = reader.getRowObjects();
-
-      return rows.map((row) => ({
-        userId: row.user_id as bigint,
-        userName: row.user_name as string,
-        fans: row.fans as number,
-        videosSeen: row.videos_seen as number,
-        videosFiltered: row.videos_filtered as number,
-        filterPassRate: row.filter_pass_rate as number,
-        discoveredFrom: row.discovered_from as "following" | "recommendation",
-        discoveredAt: new Date(row.discovered_at as string),
-        isFollowing: row.is_following as boolean,
-        lastUpdated: new Date(row.last_updated as string),
-      }));
-    });
-  }
-
-  /**
-   * Get database statistics
-   */
-  public async getStats(): Promise<DatabaseStats> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      const reader = await this.connection.runAndReadAll(`
-        SELECT 
-          (SELECT COUNT(*) FROM processed_videos) as processed_count,
-          (SELECT COUNT(*) FROM forward_dynamics) as forward_count,
-          (SELECT COUNT(*) FROM recommendations) as rec_count,
-          (SELECT COUNT(*) FROM discovered_users) as users_count,
-          (SELECT COUNT(*) FROM processed_videos WHERE is_filtered = true) as filtered_count
-      `);
-
-      const rows = reader.getRows();
-      const row = rows[0];
-
-      return {
-        processedVideosCount: row[0] as number,
-        forwardDynamicsCount: row[1] as number,
-        recommendationsCount: row[2] as number,
-        discoveredUsersCount: row[3] as number,
-        filteredVideosCount: row[4] as number,
-      };
-    });
-  }
-
-  /**
-   * Checkpoint the database to flush WAL to disk
-   */
-  public async checkpoint(): Promise<void> {
-    return this.withMutex(async () => {
-      if (!this.connection) {
-        throw new Error("Database not initialized");
-      }
-
-      await this.connection.run("CHECKPOINT");
-      logger.debug("Database checkpointed");
-    });
-  }
-
-  /**
-   * Close and reopen the database connection to reduce WAL buildup
-   */
-  public async reconnect(): Promise<void> {
-    const release = await this.mutex.acquire();
-    try {
-      if (!this.duckDBInstance) {
-        throw new Error("Database instance not initialized");
-      }
-
-      logger.debug("Reconnecting to database...");
-
-      // Close existing connection
-      if (this.connection) {
-        await this.connection.disconnect();
-        this.connection = null;
-      }
-
-      // Create new connection
-      this.connection = await this.duckDBInstance.connect();
-      logger.info("Database reconnected successfully");
-    } finally {
-      release();
-    }
-  }
-
-  /**
-   * Close the database connection
-   */
-  public async close(): Promise<void> {
-    if (this.connection) {
-      await this.connection.disconnect();
-      this.connection = null;
-    }
-    if (this.duckDBInstance) {
-      this.duckDBInstance = null;
-    }
-    logger.info("Database connection closed");
-  }
-
-  /**
-   * Get the connection (for advanced usage)
-   */
-  public getConnection(): DuckDBConnection {
-    if (!this.connection) {
-      throw new Error("Database not initialized");
-    }
-    return this.connection;
-  }
-}
